AI Finance chat/huggingface_backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-genai
python-dotenv
psycopg2-binary
transformers